chore: Change old data method
